refac: rename data_reader directory to multi_image_stack
