+++ src/careamics/dataset_ng/multi_image_stack/image_stack/__init__.py
