+++ src/careamics/dataset_ng/multi_image_stack/image_stack/in_memory_image_stack.py
