+++ src/careamics/dataset_ng/multi_image_stack/image_stack/zarr_image_stack.py
