+++ src/careamics/dataset_ng/multi_image_stack/multi_image_stack.py
